Add Main.m file via upload
Main.mlapp stores the application's code and GUI component layout - download this to open and run the Matlab App as intended. This MAIN.m version is the exact same code but does not open the App, it is a generic code file - I am uploading this for ease of viewing the code without needing to download the .mlapp file but this .m file alone will not run.